content_agregator/serializers.py
from rest_framework import serializers

from content_agregator.models import Episode


class EpisodeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Episode
        fields = '__all__'

